requirements.txt
argcomplete==1.12.3
click==8.0.1
packaging==21.0
pipx==0.16.4
pyparsing==2.4.7
python-dotenv==0.19.0
userpath==1.7.0
